requirements.txt
# Core ML frameworks
torch>=2.0.0
pytorch-lightning>=2.0.0
torchvision>=0.15.0

# Data handling
numpy>=1.21.0
scipy>=1.7.0

# Visualization and logging
tensorboard>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.64.0

# Configuration files
PyYAML>=6.0

# File handling
pathlib>=1.0.0

# Optional: Weights & Biases for logging
wandb>=0.13.0

# Optional: For advanced features
scikit-learn>=1.0.0

# Python standard library (usually included, but listing for completeness)
# typing
# warnings
# os
# json
# math
# argparse